vesuvius/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vesuvius"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.9,<3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "aiohttp>=3.12.15",
    "alphashape>=1.3.1",
    "batchgenerators>=0.25.1",
    "batchgeneratorsv2",
    "blosc2>=2.5.1",
    "build",
    "connected-components-3d>=3.24.0",
    "dask>=2024.8.0",
    "dask-image>=2024.5.3",
    "dynamic-network-architectures>=0.3.1",
    "einops>=0.8.1",
    "fft-conv-pytorch>=1.2.0",
    "fsspec>=2025.9.0",
    "fvcore>=0.1.5.post20221221",
    "huggingface-hub>=0.35.0",
    "libigl>=2.6.1",
    "lxml>=6.0.2",
    "magic-class>=0.7.17",
    "magicgui>=0.10.1",
    "monai>=1.5.1",
    "napari>=0.5.6",
    "nest-asyncio>=1.6.0",
    "nnunetv2",
    "numba>=0.60.0",
    "numcodecs>=0.12.1",
    "numpy>=2.0.2",
    "open3d",
    "open3d>=0.18,<0.19",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "psutil>=7.1.0",
    "pybind11>=3.0.1",
    "pynrrd>=1.1.3",
    "pyqt6",
    "pytest>=8.4.2",
    "pytorch-lightning>=2.5.5",
    "pytorch-optimizer>=3.8.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "s3fs>=2025.9.0",
    "scikit-image>=0.24.0",
    "scipy>=1.13.1",
    "setuptools>=80.9.0",
    "simpleitk>=2.5.2",
    "tensorstore>=0.1.69",
    "tifffile>=2024.8.30",
    "timm>=1.0.20",
    "torch>=2.6",
    "tqdm>=4.67.1",
    "trimesh>=4.8.2",
    "typed-argument-parser>=1.11.0",
    "wandb[media]>=0.22.0",
    "wheel",
    "zarr>=2,<3",
]

[project.urls]
Homepage = "https://github.com/ScrollPrize/villa"

[project.scripts]
"vesuvius.accept_terms" = "vesuvius.install.accept_terms:main"
"vesuvius.predict" = "vesuvius.models.run.inference:main"
"vesuvius.blend_logits" = "vesuvius.models.run.blending:main"
"vesuvius.finalize_outputs" = "vesuvius.models.run.finalize_outputs:main"
"vesuvius.inference_pipeline" = "vesuvius.models.run.vesuvius_pipeline:run_pipeline"
"vesuvius.compute_st" = "vesuvius.structure_tensor.run_create_st:main"
"vesuvius.napari_trainer" = "vesuvius.napari_trainer.main_window:main"
"vesuvius.voxelize_obj" = "vesuvius.image_proc.voxelize_objs:main"
"vesuvius.refine_labels" = "vesuvius.image_proc.edt_frangi_label:main"
"vesuvius.render_obj" = "vesuvius.rendering.mesh_to_surface:main"
"vesuvius.flatten_obj" = "vesuvius.rendering.slim_uv:main"
"vesuvius.train" = "vesuvius.models.training.train:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
vesuvius = ["install/configs/*.yaml"]

[tool.setuptools.cmdclass]
install = "vesuvius.install.custom_install.CustomInstallCommand"

[tool.setuptools.dynamic]
version = {attr = "vesuvius._version.VERSION"}

[tool.uv]
required-environments = [
    "python_full_version == '3.11.*' and platform_machine == 'aarch64' and sys_platform == 'linux'",
]

[tool.uv.sources]
batchgeneratorsv2 = { path = "../segmentation/models/batchgeneratorsv2" }
nnunetv2 = { path = "../segmentation/models/arch/nnunet" }
